Moved database object convertors to separate file
