+++ app/src/main/java/com/developerbreach/developerbreach/repository/database/entity/FavoritesEntity.kt
@@ -3,12 +3,11 @@ package com.developerbreach.developerbreach.repository.database.entity
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.developerbreach.developerbreach.model.Article
 import com.developerbreach.developerbreach.utils.*
 import com.developerbreach.developerbreach.utils.DatabaseProperties.ColumnFavorites
 
 @Entity(tableName = DatabaseProperties.EntityTable.FAVORITES)
-data class FavoritesEntity constructor(
+data class FavoritesEntity(
 
     @PrimaryKey
     @ColumnInfo(name = ColumnFavorites.ID)
@@ -20,21 +19,3 @@ data class FavoritesEntity constructor(
     @ColumnInfo(name = ColumnFavorites.BANNER)
     val banner: String
 )
-
-fun List<FavoritesEntity>.asDomainModel(): List<Article> {
-    return map {
-        Article(
-            articleId = it.articleId,
-            name = it.name,
-            banner = it.banner
-        )
-    }
-}
-
-fun Article.asDatabaseModel(): FavoritesEntity {
-    return FavoritesEntity(
-        articleId = this.articleId,
-        name = this.name,
-        banner = this.banner
-    )
-}
